[Gold V] Title: 2의 멱수의 합, Time: 4 ms, Memory: 5928 KB -BaekjoonHub

+++ 백준/Gold/2410. 2의 멱수의 합/2의 멱수의 합.cc
@@ -5,7 +5,7 @@
 using namespace std;
 
 int n;
-long long dp[1000001];
+int dp[1000001];
 int main()
 {
     io;
